refactor hello-world into tag search service
