--- backend-app/.aws-sam/build.toml
@@ -1,11 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions]
-[function_build_definitions.5d170682-7502-4ed8-8f73-7389c2be84c8]
-codeuri = "/Users/sanderdemeester/Documents/coding/get-hashdag/backend-app/hello-world"
-runtime = "go1.x"
-source_md5 = ""
-packagetype = "Zip"
-functions = ["HelloWorldFunction"]
-
-[layer_build_definitions]
