+++ backend-app/tag-search-api/go.mod
@@ -1,0 +1,5 @@
+require github.com/aws/aws-lambda-go v1.23.0
+
+module hello-world
+
+go 1.16
