+++ backend-app/tag-search-api/request_router.go
@@ -1,0 +1,7 @@
+package main
+
+import "fmt"
+
+func SimpleRouter(){
+	fmt.Println("SimpleRouter")
+}
