+++ backend-app/tag-search-api/storage_service.go
@@ -1,0 +1,7 @@
+package main
+
+import "fmt"
+
+func SimpleDynamoDBQuery() {
+	fmt.Println("SimpleDynamoDBQuery")
+}
