+++ hashdag.code-workspace
@@ -8,20 +8,37 @@
 		"configurations": [{
 			"type": "aws-sam",
 			"request": "direct-invoke",
-			"name": "API hashdag:helloworld"
+			"name": "API hashdag:results",
 			"invokeTarget":{
 				"target": "api",
 				"templatePath": "${workspaceFolder}/backend-app/template.yaml",
-				"logicalId": "HelloWorldFunction",
+				"logicalId": "TagSearchApiFunction",
 			},
 			"api":{
-				"path":"/hello",
+				"path":"/results",
 				"httpMethod": "get",
 				"payload": {
 					"json":{}
 				}
 			}
-
+		},{
+			"type": "aws-sam",
+			"request": "direct-invoke",
+			"name": "API hashdag:register",
+			"invokeTarget":{
+				"target": "api",
+				"templatePath": "${workspaceFolder}/backend-app/template.yaml",
+				"logicalId": "TagSearchApiFunction",
+			},
+			"api":{
+				"path":"/register",
+				"httpMethod": "post",
+				"payload": {
+					"json":{
+						"searchtag": "Vancouver",
+					}
+				}
+			}
 		}],
 		"compounds": []
 	}
